+++ NEWS.md
@@ -1,7 +1,21 @@
-# RstoxData v1.11.3  (2023-06-27)
+# RstoxBase v1.11.4-9001  (2023-08-30)
+* Removed dependency on the retiring package sp.
+* Changed GearDependentCatchCompensation() to keep all variables from the input SpeciesCategoryCatchData.
+* Improved error message then Percentages is not given (now saying exactly that and not "SpecificationParameter must be given").
+* Removed exported functions getStratumPolygonList and readGeoJSON.
+* Improved documentation of DefinitionMethod in DefineBioticPSU(), DefineAcousticPSU() and DefineBioticAssignment().
+* Added a warning when reading BioticPSUs from a StoX 2.7 project.xml file where Method is Station and not UseProcessData in DefineSweptAreaPSU(), which makes the BioticPSUs of the project.xml file unused.
+* Improved warning when there are Individuals in the IndividualsData with IndividualTotalLength that does not match any of the length intervals of the QuantityData.
+* Improved warning for when there are positive NASC values with no assignment length distribution, also removing the list of the  affected PSUs.
+* Improved simplifyStratumPolygon() used in DefineStratumPolygon() which got stuck in an endless loop in certain cases.
+* Exposing PlotAcousticTrawlSurvey().
+* Fixed bug in LengthDistribution() where missing raising factor was reported for samples with no individuals.
+
+
+# RstoxBase v1.11.3  (2023-06-27)
 * Version for StoX 3.6.2.
 * Changed defined unit of Biomass in QuantityData from g to kg, as Biomass originates from SweptAreaMethod "TotalCatch" which is in kg.
-* Added tables listing the variables of the data types NASC, SumNASC, MeanNASC, LengthDistribution, SumLengthDistribution, MeanLengthDistribution, Density, MeanDensity and Quantity. 
+* Added tables listing the variables of the data types NASCData, SumNASCData, MeanNASCData, LengthDistributionData, SumLengthDistributionData, MeanLengthDistributionData, DensityData, MeanDensityData and QuantityData.  
 * Relaxed testUnits.R to accept that Biomass has different units in QuantityData and SuperIndividualsData.
 * Increased speed of frequently used functions in RstoxFramework by using list instead of data.table for the definition og reportFunctions.
 * Skipped gc() in aggregateBaselineDataOneTable() to reduce CPU time.
@@ -314,7 +328,7 @@
 * Fixed bug in SplitMeanNASC(), where now acoustic categories not to be split are added to the AcousticCategoryLink only if of positive length.
 * Fixed warning 'the condition has length > 1 and only the first element will be used'
 * Fixed bug with multiple species in catchCompensationTable, selectivityTable and gearCompensationTable parameter formats.
-* Added stop if not all EDSUs are inside a stratum in SplitMeanNASC()Fixed bug where stations with 0 fish were removed in LenghtDistribution()
+* Added stop if not all EDSUs are inside a stratum in SplitMeanNASC()Fixed bug where stations with 0 fish were removed in LengthDistribution()
 * Fixed several bugs related to DensityData and onwards being flexible datatypes which may or may not contain Beam and Frequency
 * Added parameters to ImputeSuperIndividuals().
 * Fixed column order for datatypes *IndividualsData